feat(2D-DOC-Parsing): Base code for data extraction
This code handles the data extraction.

For now only supports the extraction of the document 01 "Justificatif de domicile"
This document was used because it covers all cases for the data formats inside the 2D-DOC :
- Fixed length data
- Variable length data with a max length
- Variable length data with no max length

The other document types will be added later.

This code also supports the parsing of the extracted data. Each data type is of some data format and those data format needs to be parsed into the right Java type for the library user to use.